# docs/swagger.yaml
basePath: /api
definitions:
  handlers.BorrowerResponse:
    description: Response containing borrower data
    properties:
      contact_info:
        type: string
      id:
        type: string
      is_delinquent:
        type: boolean
      name:
        type: string
    type: object
  handlers.CreateBorrowerRequest:
    description: Request body for creating a new borrower
    properties:
      contact_info:
        type: string
      name:
        type: string
    required:
    - contact_info
    - name
    type: object
  handlers.CreateLoanRequest:
    description: Request body for creating a new loan
    properties:
      amount:
        minimum: 1
        type: integer
      borrower_id:
        type: string
      interest_rate:
        minimum: 0
        type: number
      term_weeks:
        minimum: 1
        type: integer
    required:
    - amount
    - borrower_id
    - interest_rate
    - term_weeks
    type: object
  handlers.LoanResponse:
    description: Response containing loan data
    properties:
      amount:
        type: integer
      borrower_id:
        type: string
      id:
        type: string
      interest_rate:
        type: number
      start_date:
        type: string
      status:
        type: string
      term_weeks:
        type: integer
    type: object
  handlers.PaymentRequest:
    description: Request body for making a payment
    properties:
      amount:
        minimum: 1
        type: integer
    required:
    - amount
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing loans, borrowers, payments, and delinquency status
  title: Loan Billing System API
  version: "1.0"
paths:
  /api/borrowers:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all borrowers in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BorrowerResponse'
            type: array
        "500":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all borrowers
      tags:
      - Borrowers
    post:
      consumes:
      - application/json
      description: Creates a new borrower in the system
      parameters:
      - description: Borrower details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBorrowerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.BorrowerResponse'
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new borrower
      tags:
      - Borrowers
  /api/borrowers/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves details for a specific borrower
      parameters:
      - description: Borrower ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BorrowerResponse'
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get borrower details
      tags:
      - Borrowers
  /api/borrowers/delinquent:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all borrowers who are currently delinquent
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BorrowerResponse'
            type: array
        "500":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List delinquent borrowers
      tags:
      - Borrowers
  /api/loans:
    post:
      consumes:
      - application/json
      description: Creates a new loan for a borrower
      parameters:
      - description: Loan details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateLoanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.LoanResponse'
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new loan
      tags:
      - Loans
  /api/loans/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves details for a specific loan
      parameters:
      - description: Loan ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoanResponse'
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get loan details
      tags:
      - Loans
  /api/loans/{id}/delinquent:
    get:
      consumes:
      - application/json
      description: Checks if a loan is currently delinquent (2+ missed payments)
      parameters:
      - description: Loan ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delinquency status
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check if loan is delinquent
      tags:
      - Loans
  /api/loans/{id}/outstanding:
    get:
      consumes:
      - application/json
      description: Retrieves the current outstanding balance for a loan
      parameters:
      - description: Loan ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Outstanding amount
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get outstanding balance
      tags:
      - Loans
  /api/loans/{id}/payment:
    post:
      consumes:
      - application/json
      description: Makes a payment for a loan
      parameters:
      - description: Loan ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Payment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Make a payment
      tags:
      - Payments
swagger: "2.0"